Regression fix: Sunlight sprite
Turns out this has been broken for a while. I had to dig fairly deep
into my archives to find a copy that wasn't broken. Then I just updated
it to work with the new system.

Also got rid of the weird OverlayPixmaps function.
And fixed a bug with the Palm Tree.
And a few other random things.