Fix a game patch bug, and add debug stuff
Fixed a bug that can cause the Zones dialog to crash sometimes, among
other things. Also added a debugger for a mysterious bug that seems to
happen sporadically. If it happens again, the new ValueError will show
almost all of the values passed to struct.pack_into(), which is where
the bug shows up. If somebody gets the error again and reports the error
log, the bug can be diagnosed properly. (Rationale: I don't know which
argument is sometimes not within 0-255. This will help.)