Add lots of abstraction
If this is ever going to work, Reggie will need a lot more abstraction.
I've added abstraction to both the mainWindow.LoadLevel() function and
the level class.

Side note: Don't try to save NSMB2 levels yet. It will "work", but it
saves in the NSMBW level format. This can be opened by normal Reggie,
but NSMB2 Reggie will crash badly when re-opening it, eating up your RAM
in the process.