SARC Hash Multiplier
It used to be an unknown value, but now it's not. So I changed the code
to reflect that. You can also set your own hash multiplier values if you
want. Nintendo seems to always use 0x65, and there's no reason to use
anything else, but it's nice to have options. :P