Add level analysis tool and path support
"Level analysis tool": meant for programmers, not average users.

It's a loop that opens every NSMB2 level. You can then add code to other
parts of the editor to gather information. Tip: use the new global
variable "CurrentLevelNameForAutoOpenScript", which (when the
auto-opening loop is enabled) will always contain the name of the level
currently being opened.
You need to change the AutoOpenScriptEnabled variable to True within the
source code for this to work.

Path support: Reggie now supports NSMB2 paths. That's about it.

There are also a few minor bug fixes included.