Make zone entrance marker lines optional
It's turned off by default. The setting is in the Preferences dialog, if
you want to turn it on. I also made it a bit smarter - if the option in
Preferences is turned on, the lines only display if the Zone Direction
option is set to "Left to Right" or "Right to Left".